+++ RELEASE_NOTES.rst
@@ -9,6 +9,7 @@ All changes
 - Update reference for activity and capacity soft constraints (:pull:`474`).
 - Update :meth:`.years_active` to use sorted results (:pull:`491`).
 - Adjust the Westeros reporting tutorial to pyam 1.0 deprecations (:pull:`492`).
+- Change precision of GAMS check for parameter "duration_time" (:pull:`513`).
 
 
 .. _v3.3.0: